requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv

# Project requirements
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken

# uncomment to use huggingface llms
huggingface-hub

# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers
